+++ player_control.h
@@ -180,7 +180,7 @@ class PlayerControl : public CreatureView, public CollectiveControl {
   void handleMarket(View*, int prevItem = 0);
   void getEquipmentItem(View* view, ItemPredicate predicate);
   Item* chooseEquipmentItem(View* view, vector<Item*> currentItems, ItemPredicate predicate,
-      int* index = nullptr, double* scrollPos = nullptr) const;
+      int* index = nullptr, int* scrollPos = nullptr) const;
 
   void getMinionOptions(Creature*, vector<MinionOption>&, vector<View::ListElem>&);
